refactor: move mockLocalStorage() to testutil
