+++ src/app/features/operation-table/operation-table-view/operation-table-view.component.spec.ts
@@ -1,13 +1,13 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { OperationTableView } from './operation-table-view.component';
-import { mockLocalStorage } from 'src/app/core/util/testing';
 import { IncidentService } from 'src/app/core/services/incident/incident.service';
 import { LocalStorageIncidentService } from 'src/app/core/services/incident/local-storage-incident.service';
 import { NotificationService } from 'src/app/core/services/notification.service';
 import { LocalStorageResourceService } from 'src/app/core/services/resource/local-storage-resource.service';
 import { ResourceService } from 'src/app/core/services/resource/resource.service';
 import { MAT_DIALOG_DATA, MatDialogModule, MatDialogRef } from '@angular/material/dialog';
+import { mockLocalStorage } from 'src/app/core/testutil/testutil';
 
 describe('OperationTableView', () => {
   let component: OperationTableView;
